Revert "feature: reduce memory footprint when storing SD template mapping (#1…"
This reverts commit 53fbe956e77a64779a8c2f546205fbce96e14c5e.
