+++ src/main/kotlin/net/atos/zac/smartdocuments/templates/model/SmartDocumentsTemplateGroup.kt
@@ -17,7 +17,7 @@ import jakarta.persistence.ManyToOne
 import jakarta.persistence.OneToMany
 import jakarta.persistence.SequenceGenerator
 import jakarta.persistence.Table
-import net.atos.zac.admin.model.ZaakafhandelParametersSummary
+import net.atos.zac.admin.model.ZaakafhandelParameters
 import net.atos.zac.util.FlywayIntegrator
 import nl.lifely.zac.util.AllOpen
 import java.time.ZonedDateTime
@@ -58,5 +58,5 @@ class SmartDocumentsTemplateGroup {
 
     @ManyToOne
     @JoinColumn(name = "zaakafhandelparameters_id", nullable = false)
-    lateinit var zaakafhandelParameters: ZaakafhandelParametersSummary
+    lateinit var zaakafhandelParameters: ZaakafhandelParameters
 }
